refactor:(authorization-service): remove overcomplication, use firebase directly (#16)
The firebase abstraction layer was pointless and heavy.
Adds some better JSON handling, to help prevent issues that arise as the model objects get
heavier and harder to manually keep the inputs and outputs in sync during development.

More tests.

BREAKING CHANGE:

role and permission getters on the User and visitor services are now Promises. Using
Observables was a lie, and overcomplicated the code.