feat(all): Improvements from development of https://SnapLog.io
This (massive, breaks-everything) change set brings in the changes to
Tangential that were made during, and informed by, the development of
https://SnapLog.io.

If you are developing against a clone of this project you should probably
not attempt a merge, as the changes are so widespread as to make a successful
merge incredibly unlikely. Consider instead pulling back in
and changes you've made on top of a clean clone.

Documentation will be updated shortly.

We will be updating the layout of the auth document in the firebase database
in a coming change, in order to support pulling down all the roles and
permissions in one request, while not consuming the (potentially vast) users
document. We will also be moving the user activity log into a separate
document, to help keep the size of each user consistent, etc.

BREAKING CHANGE: Basically everything. Please consider this a clean start. We
won't be making such large changes in the future. Promise.

BREAKING CHANGE: The tg- prefix is now tanj-. This should be amenable to
global search and replace.