feat(firebase-functions): Add watch for setting new user Role
New users are granted the default role specified in
`auth/settings/configuration/`. Anonymous accounts and regular
accounts receive different roles.

Updates the default database init file with a new permission for
viewing users, and new default permissions for the
'Registered Visitor' Role.

WARNING: There is not yet a watch that will update a user's role
from the 'defaultAnonymousRole' to the 'defaultUserRole' value when
the user registers and links to an existing anonymous acount.