refactor(authorization-service): Cleanup the visitor/subject paradigm
Define the heirarchy as User -> Subject -> Visitor

Visitor will be renmaed 'CurrentVisitor' at some point in the future,
and an intermediary class (e.g. AppUser) created, so that we can
manage our Application's users, including their data.

This change doesn't fix the tests it broke, that's for next commit.
The testing of the auth service is going to go to a Composite Model
focus - there's no point in testing things we never do, like
grabbing only the rules or only the rulePermission mappings.