feat(core): Logger should be injected, not static
The logging system was not quite as nice as it should have been.
Loggers are now injected, and called via 'this'. If you don't want any
logging at all  you should implement a NullLogger or similar, that extends
the abstract Logger class and does nothing.

This change set also improves the BusMessage class to clarify expected
usage patterns.